Fix renaming things and whatnot
